muse_demo/requirements.txt
streamlit>=1.32.0
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0
onnxruntime>=1.17.0
numpy>=1.24.0
scipy>=1.11.0
plotly>=5.18.0
pyyaml>=6.0
